minor polishing
